+++ packages/dev/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gestaltjs/dev
 
+## 0.7.7
+
+### Patch Changes
+
+- @gestaltjs/core@0.7.7
+
 ## 0.7.6
 
 ### Patch Changes