+++ packages/plugin-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gestaltjs/gestalt-plugin-react
 
+## 0.7.7
+
+### Patch Changes
+
+- Make "@gestaltjs/plugins" a dependency of the renderer packages
+  - @gestaltjs/plugins@0.7.7
+
 ## 0.7.6
 
 ### Patch Changes